fixup! offers: add merkle selective-disclosure primitives
