+++ lightning/src/offers/mod.rs
@@ -27,6 +27,7 @@ pub mod parse;
 mod payer;
 pub mod payer_proof;
 pub mod refund;
+pub mod selective_disclosure;
 pub(crate) mod signer;
 pub mod static_invoice;
 #[cfg(test)]
